refactor(sb): timetable csf3
